requirements.txt
requests
python-dotenv
pywin32
supabase
pypdf